pyproject.toml
[project]
name = "snapshot"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Edgar E. Reyes T.", email = "linuosx@gmail.com"},
]
dependencies = [
    "azure-identity",
    "azure-mgmt-compute",
    "python-decouple",
]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"
